Move outbound peer code out from base class
