+++ elements/lisk-p2p/src/peer/outbound.ts
@@ -14,31 +14,48 @@
  */
 
 import {
+	FetchPeerStatusError,
+	PeerOutboundConnectionError,
+	RPCResponseError,
+} from '../errors';
+
+import {
 	ClientOptionsUpdated,
+	ConnectAndFetchResponse,
 	convertNodeInfoToLegacyFormat,
 	DEFAULT_ACK_TIMEOUT,
 	DEFAULT_CONNECT_TIMEOUT,
-	EVENT_CLOSE_OUTBOUND,
-	EVENT_CONNECT_ABORT_OUTBOUND,
-	EVENT_CONNECT_OUTBOUND,
-	EVENT_OUTBOUND_SOCKET_ERROR,
 	Peer,
 	PeerConfig,
 	REMOTE_EVENT_MESSAGE,
 	REMOTE_EVENT_RPC_REQUEST,
+	REMOTE_RPC_GET_NODE_INFO,
 } from './base';
 
 import {
 	P2PDiscoveredPeerInfo,
 	P2PMessagePacket,
+	P2PNodeInfo,
+	P2PPeerInfo,
 	P2PRequestPacket,
 	P2PResponsePacket,
 } from '../p2p_types';
 
 import * as querystring from 'querystring';
 import * as socketClusterClient from 'socketcluster-client';
+import { validatePeerInfo } from '../validation';
 
 type SCClientSocket = socketClusterClient.SCClientSocket;
+
+export const EVENT_CONNECT_OUTBOUND = 'connectOutbound';
+export const EVENT_CONNECT_ABORT_OUTBOUND = 'connectAbortOutbound';
+export const EVENT_CLOSE_OUTBOUND = 'closeOutbound';
+export const EVENT_OUTBOUND_SOCKET_ERROR = 'outboundSocketError';
+
+export interface PeerInfoAndOutboundConnection {
+	readonly peerInfo: P2PDiscoveredPeerInfo;
+	readonly socket: SCClientSocket;
+}
 
 export class OutboundPeer extends Peer {
 	protected _socket: SCClientSocket | undefined;
@@ -189,3 +206,134 @@ export class OutboundPeer extends Peer {
 		);
 	}
 }
+
+export const connectAndRequest = async (
+	basicPeerInfo: P2PPeerInfo,
+	procedure: string,
+	nodeInfo?: P2PNodeInfo,
+	peerConfig?: PeerConfig,
+): Promise<ConnectAndFetchResponse> =>
+	new Promise<ConnectAndFetchResponse>(
+		(resolve, reject): void => {
+			const legacyNodeInfo = nodeInfo
+				? convertNodeInfoToLegacyFormat(nodeInfo)
+				: undefined;
+			// Add a new field discovery to tell the receiving side that the connection will be short lived
+			const requestPacket = {
+				procedure,
+			};
+			// Ideally, we should JSON-serialize the whole NodeInfo object but this cannot be done for compatibility reasons, so instead we put it inside an options property.
+			const clientOptions: ClientOptionsUpdated = {
+				hostname: basicPeerInfo.ipAddress,
+				port: basicPeerInfo.wsPort,
+				query: querystring.stringify({
+					...legacyNodeInfo,
+					options: JSON.stringify(legacyNodeInfo),
+				}),
+				connectTimeout: peerConfig
+					? peerConfig.connectTimeout
+						? peerConfig.connectTimeout
+						: DEFAULT_CONNECT_TIMEOUT
+					: DEFAULT_CONNECT_TIMEOUT,
+				ackTimeout: peerConfig
+					? peerConfig.connectTimeout
+						? peerConfig.connectTimeout
+						: DEFAULT_CONNECT_TIMEOUT
+					: DEFAULT_ACK_TIMEOUT,
+				multiplex: false,
+				autoConnect: false,
+				autoReconnect: false,
+				pingTimeoutDisabled: true,
+			};
+
+			const outboundSocket = socketClusterClient.create(clientOptions);
+			// Bind an error handler immediately after creating the socket; otherwise errors may crash the process
+			// tslint:disable-next-line no-empty
+			outboundSocket.on('error', () => {});
+
+			// tslint:disable-next-line no-let
+			let disconnectStatusCode: number;
+			// tslint:disable-next-line no-let
+			let disconnectReason: string;
+			const closeHandler = (statusCode: number, reason: string) => {
+				disconnectStatusCode = statusCode;
+				disconnectReason = reason;
+			};
+			outboundSocket.once('close', closeHandler);
+
+			// Attaching handlers for various events that could be used future for logging or any other application
+			outboundSocket.emit(
+				REMOTE_EVENT_RPC_REQUEST,
+				{
+					type: '/RPCRequest',
+					procedure: requestPacket.procedure,
+				},
+				(err: Error | undefined, responseData: unknown) => {
+					outboundSocket.off('close', closeHandler);
+					if (err) {
+						const isFailedConnection =
+							disconnectReason &&
+							(err.name === 'TimeoutError' ||
+								err.name === 'BadConnectionError');
+						const connectionError = new PeerOutboundConnectionError(
+							isFailedConnection ? disconnectReason : err.message,
+							disconnectStatusCode,
+						);
+						reject(connectionError);
+
+						return;
+					}
+					if (responseData) {
+						const responsePacket = responseData as P2PResponsePacket;
+						resolve({
+							responsePacket,
+							socket: outboundSocket,
+						});
+
+						return;
+					}
+
+					reject(
+						new RPCResponseError(
+							`Failed to handle response for procedure ${
+								requestPacket.procedure
+							}`,
+							basicPeerInfo.ipAddress,
+							basicPeerInfo.wsPort,
+						),
+					);
+				},
+			);
+		},
+	);
+
+export const connectAndFetchPeerInfo = async (
+	basicPeerInfo: P2PPeerInfo,
+	nodeInfo?: P2PNodeInfo,
+	peerConfig?: PeerConfig,
+): Promise<PeerInfoAndOutboundConnection> => {
+	try {
+		const { responsePacket, socket } = await connectAndRequest(
+			basicPeerInfo,
+			REMOTE_RPC_GET_NODE_INFO,
+			nodeInfo,
+			peerConfig,
+		);
+
+		const protocolPeerInfo = responsePacket.data;
+		const rawPeerInfo = {
+			...protocolPeerInfo,
+			ip: basicPeerInfo.ipAddress,
+			wsPort: basicPeerInfo.wsPort,
+		};
+		const peerInfo = validatePeerInfo(rawPeerInfo);
+
+		return { peerInfo, socket };
+	} catch (error) {
+		throw new FetchPeerStatusError(
+			`Error occurred while fetching information from ${
+				basicPeerInfo.ipAddress
+			}:${basicPeerInfo.wsPort}`,
+		);
+	}
+};
